backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
elasticsearch[async]>=8.12.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
email-validator>=2.0.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.27.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
